working recv game packets

+++ KastiBob/logic/src/packets/LoginPacket.hpp
@@ -16,7 +16,7 @@ public:
 	LoginPacket(std::string login, std::string password,
 		uint16_t tibiaversion, uint16_t ostype, uint32_t dat_signature,
 		uint32_t spr_signature, uint32_t pic_signature, 
-		uint32_t data[5], std::string name,
+		uint8_t data[5], std::string name,
 		RSAcipher& rsa, XTEAcipher& xtea);
 };
 #endif
